Rename files and directories for consistency
