+++ src/components/HomePage/GameRunning/GameRunning.scss
