+++ src/components/HomePage/GameRunning/index.js
