+++ src/components/HomePage/StartGame/StartGame.scss
