+++ src/components/HomePage/StartGame/index.js
