+++ src/components/Timer/Timer.scss
